Removed a large majority of typescript code and replaced with JS/ES6
