+++ lib/package.json
@@ -7,7 +7,6 @@
   "browser": {
     "./dist/vue-easy-dnd.ssr.js": "./dist/vue-easy-dnd.js"
   },
-  "types": "dist/index.d.ts",
   "files": [
     "dist",
     "src"
