+++ lib/rollup.config.js
@@ -17,7 +17,7 @@ const name = pkg.name;
 // de la sorte. En général, cette liste est déduite des dépendances dans package.json :
 const external = Object.keys(pkg.dependencies || {});
 
-const input = path.resolve(__dirname, './src/index.ts');
+const input = path.resolve(__dirname, './src/index.js');
 
 const plugins = [
     commonjs(),
@@ -65,9 +65,6 @@ export default [
         output: {
             format: 'iife',
             name: "VueEasyDnD",
-            globals: {
-                'vue-property-decorator': 'VuePropertyDecorator',
-            },
             file: path.resolve(__dirname, './dist/' + name + '.js'),
         }
     }
