+++ lib/src/ts/Grid.js
@@ -5,7 +5,7 @@ export default class Grid {
     magnets: { x, y }[] = [];
 
     constructor(collection: HTMLCollection, upToIndex: number, row: string, fromIndex: number) {
-        this.reference = collection.item(0).parentNode as HTMLElement;
+        this.reference = collection.item(0).parentNode; // todo - as HTMLElement was removed (TS)
         this.referenceOriginalPosition = {
             x: this.reference.getBoundingClientRect().left - this.reference.scrollLeft,
             y: this.reference.getBoundingClientRect().top - this.reference.scrollTop,
