+++ lib/src/ts/createDragImage.js
@@ -20,13 +20,13 @@ export function createDragImage(el: HTMLElement): HTMLElement {
  * Clones the given element and all its descendants.
  */
 function deepClone(el: HTMLElement): HTMLElement {
-    let clone = el.cloneNode(true) as HTMLElement;
+    let clone = el.cloneNode(true); // todo - as HTMLElement
     copyStyle(el, clone);
     let vSrcElements = el.getElementsByTagName("*");
     let vDstElements = clone.getElementsByTagName("*");
     for (let i = vSrcElements.length; i--;) {
-        let vSrcElement = vSrcElements[i] as HTMLElement;
-        let vDstElement = vDstElements[i] as HTMLElement;
+        let vSrcElement = vSrcElements[i]; // todo - as HTMLElement
+        let vDstElement = vDstElements[i]; // todo - as HTMLElement
         copyStyle(vSrcElement, vDstElement);
     }
     return clone;
