+++ src/App3.vue
@@ -65,69 +65,68 @@
 </template>
 
 <script>
-    import Drag from "../lib/src/components/Drag.vue";
-    import Drop from "../lib/src/components/Drop.vue";
-    import DropList from "../lib/src/components/DropList.vue";
-    import "../lib/src/ts/DragImagesManager.ts";
+import Drag from "../lib/src/components/Drag";
+import Drop from "../lib/src/components/Drop";
+import DropList from "../lib/src/components/DropList";
 
-    export default {
-        name: "App",
-        components: {
-            Drag,
-            DropList,
-            Drop
+export default {
+    name: "App",
+    components: {
+        Drag,
+        DropList,
+        Drop
+    },
+    data: function () {
+        return {
+            items1: [
+                {
+                    avatar: "https://cdn.vuetifyjs.com/images/lists/1.jpg",
+                    title: "Brunch this weekend?",
+                    subtitle:
+                        "<span class='text--primary'>Ali Connors</span> &mdash; I'll be in your neighborhood doing errands this weekend. Do you want to hang out?"
+                },
+                {
+                    avatar: "https://cdn.vuetifyjs.com/images/lists/2.jpg",
+                    title: "Summer BBQ",
+                    subtitle:
+                        "<span class='text--primary'>to Alex, Scott, Jennifer</span> &mdash; Wish I could come, but I'm out of town this weekend."
+                },
+                {
+                    avatar: "https://cdn.vuetifyjs.com/images/lists/3.jpg",
+                    title: "Oui oui",
+                    subtitle:
+                        "<span class='text--primary'>Sandra Adams</span> &mdash; Do you have Paris recommendations? Have you ever been?"
+                }
+            ],
+            items2: [
+                {
+                    avatar: "https://cdn.vuetifyjs.com/images/lists/4.jpg",
+                    title: "Birthday gift",
+                    subtitle:
+                        "<span class='text--primary'>Trevor Hansen</span> &mdash; Have any ideas about what we should get Heidi for her birthday?"
+                },
+                {
+                    avatar: "https://cdn.vuetifyjs.com/images/lists/5.jpg",
+                    title: "Recipe to try",
+                    subtitle:
+                        "<span class='text--primary'>Britta Holt</span> &mdash; We should eat this: Grate, Squash, Corn, and tomatillo Tacos."
+                }
+            ]
+        };
+    },
+    methods: {
+        insert1(event) {
+            this.items1.splice(event.index, 0, event.data);
         },
-        data: function () {
-            return {
-                items1: [
-                    {
-                        avatar: "https://cdn.vuetifyjs.com/images/lists/1.jpg",
-                        title: "Brunch this weekend?",
-                        subtitle:
-                            "<span class='text--primary'>Ali Connors</span> &mdash; I'll be in your neighborhood doing errands this weekend. Do you want to hang out?"
-                    },
-                    {
-                        avatar: "https://cdn.vuetifyjs.com/images/lists/2.jpg",
-                        title: "Summer BBQ",
-                        subtitle:
-                            "<span class='text--primary'>to Alex, Scott, Jennifer</span> &mdash; Wish I could come, but I'm out of town this weekend."
-                    },
-                    {
-                        avatar: "https://cdn.vuetifyjs.com/images/lists/3.jpg",
-                        title: "Oui oui",
-                        subtitle:
-                            "<span class='text--primary'>Sandra Adams</span> &mdash; Do you have Paris recommendations? Have you ever been?"
-                    }
-                ],
-                items2: [
-                    {
-                        avatar: "https://cdn.vuetifyjs.com/images/lists/4.jpg",
-                        title: "Birthday gift",
-                        subtitle:
-                            "<span class='text--primary'>Trevor Hansen</span> &mdash; Have any ideas about what we should get Heidi for her birthday?"
-                    },
-                    {
-                        avatar: "https://cdn.vuetifyjs.com/images/lists/5.jpg",
-                        title: "Recipe to try",
-                        subtitle:
-                            "<span class='text--primary'>Britta Holt</span> &mdash; We should eat this: Grate, Squash, Corn, and tomatillo Tacos."
-                    }
-                ]
-            };
+        insert2(event) {
+            this.items2.splice(event.index, 0, event.data);
         },
-        methods: {
-            insert1(event) {
-                this.items1.splice(event.index, 0, event.data);
-            },
-            insert2(event) {
-                this.items2.splice(event.index, 0, event.data);
-            },
-            remove(array, value) {
-                let index = array.indexOf(value);
-                array.splice(index, 1);
-            }
+        remove(array, value) {
+            let index = array.indexOf(value);
+            array.splice(index, 1);
         }
-    };
+    }
+};
 </script>
 
 <style>
