+++ src/App4.vue
@@ -47,87 +47,86 @@
 </template>
 
 <script>
-    import Drag from "../lib/src/components/Drag.vue";
-    import Drop from "../lib/src/components/Drop.vue";
-    import DropList from "../lib/src/components/DropList.vue";
-    import "../lib/src/ts/DragImagesManager.ts";
+import Drag from "../lib/src/components/Drag";
+import Drop from "../lib/src/components/Drop";
+import DropList from "../lib/src/components/DropList";
 
-    export default {
-        name: "App",
-        components: {
-            Drag,
-            DropList,
-            Drop
-        },
-        data: function () {
-            return {
-                headers: [
-                    {text: 'Dessert (100g serving)', value: 'name'},
-                    {text: 'Calories', value: 'calories'},
-                    {text: 'Fat (g)', value: 'fat'}
-                ],
-                desserts: [
-                    {
-                        name: 'Frozen Yogurt',
-                        calories: 159,
-                        fat: 6.0
-                    },
-                    {
-                        name: 'Ice cream sandwich',
-                        calories: 237,
-                        fat: 9.0
-                    },
-                    {
-                        name: 'Eclair',
-                        calories: 262,
-                        fat: 16.0
-                    },
-                    {
-                        name: 'Cupcake',
-                        calories: 305,
-                        fat: 3.7
-                    },
-                    {
-                        name: 'Gingerbread',
-                        calories: 356,
-                        fat: 16.0
-                    },
-                    {
-                        name: 'Jelly bean',
-                        calories: 375,
-                        fat: 0.0
-                    },
-                    {
-                        name: 'Lollipop',
-                        calories: 392,
-                        fat: 0.2
-                    },
-                    {
-                        name: 'Honeycomb',
-                        calories: 408,
-                        fat: 3.2
-                    },
-                    {
-                        name: 'Donut',
-                        calories: 452,
-                        fat: 25.0
-                    },
-                    {
-                        name: 'KitKat',
-                        calories: 518,
-                        fat: 26.0
-                    },
-                ],
-                selected: [],
-                lists: []
-            };
-        },
-        methods: {
-            onDrop(event) {
-                this.lists.push(event.data);
-            }
+export default {
+    name: "App",
+    components: {
+        Drag,
+        DropList,
+        Drop
+    },
+    data: function () {
+        return {
+            headers: [
+                {text: 'Dessert (100g serving)', value: 'name'},
+                {text: 'Calories', value: 'calories'},
+                {text: 'Fat (g)', value: 'fat'}
+            ],
+            desserts: [
+                {
+                    name: 'Frozen Yogurt',
+                    calories: 159,
+                    fat: 6.0
+                },
+                {
+                    name: 'Ice cream sandwich',
+                    calories: 237,
+                    fat: 9.0
+                },
+                {
+                    name: 'Eclair',
+                    calories: 262,
+                    fat: 16.0
+                },
+                {
+                    name: 'Cupcake',
+                    calories: 305,
+                    fat: 3.7
+                },
+                {
+                    name: 'Gingerbread',
+                    calories: 356,
+                    fat: 16.0
+                },
+                {
+                    name: 'Jelly bean',
+                    calories: 375,
+                    fat: 0.0
+                },
+                {
+                    name: 'Lollipop',
+                    calories: 392,
+                    fat: 0.2
+                },
+                {
+                    name: 'Honeycomb',
+                    calories: 408,
+                    fat: 3.2
+                },
+                {
+                    name: 'Donut',
+                    calories: 452,
+                    fat: 25.0
+                },
+                {
+                    name: 'KitKat',
+                    calories: 518,
+                    fat: 26.0
+                },
+            ],
+            selected: [],
+            lists: []
+        };
+    },
+    methods: {
+        onDrop(event) {
+            this.lists.push(event.data);
         }
-    };
+    }
+};
 </script>
 
 <style>
