+++ src/components/App12Item.vue
@@ -14,7 +14,6 @@
 
 <script>
 export default {
-
 }
 </script>
