+++ src/components/Column.vue
@@ -2,22 +2,12 @@
     <flex :items="items" direction="column"></flex>
 </template>
 
-<script lang="ts">
-    import {Component, Prop, Vue} from "vue-property-decorator";
-    import Flex from "@/components/Flex.vue";
+<script>
+import Flex from "./Flex";
 
-    @Component({
-        name: 'Column',
-        components: {Flex}
-    })
-    export default class Column extends Vue {
-
-        @Prop()
-        items: any[];
-
-    }
+export default {
+  name: 'Column',
+  components: { Flex },
+  props: ['items']
+}
 </script>
-
-<style scoped lang="scss">
-
-</style>
