+++ src/components/MyDiv.vue
@@ -4,14 +4,8 @@
     </div>
 </template>
 
-<script lang="ts">
-    import {Component, Vue} from "vue-property-decorator";
-
-    @Component({})
-    export default class MyDiv extends Vue {
-    }
+<script>
+export default {
+    name: 'MyDiv'
+}
 </script>
-
-<style scoped lang="scss">
-
-</style>
