+++ src/components/Row.vue
@@ -2,22 +2,12 @@
     <flex :items="items" direction="row"></flex>
 </template>
 
-<script lang="ts">
-    import {Component, Prop, Vue} from "vue-property-decorator";
-    import Flex from "@/components/Flex.vue";
+<script>
+import Flex from "./Flex.vue";
 
-    @Component({
-        name: 'Row',
-        components: {Flex}
-    })
-    export default class Row extends Vue {
-
-        @Prop()
-        items: any[];
-
-    }
+export default {
+  name: 'Row',
+  components: { Flex },
+  props: ['items']
+}
 </script>
-
-<style scoped lang="scss">
-
-</style>
